하원 후 집에서 작업 멤버 디테일 js, server작업중

+++ sixman/src/main/java/com/kh/sixman/file/controller/FileController.java
@@ -49,9 +49,10 @@ public class FileController {
 	@PostMapping("delete")
 	public void delete(@RequestParam Map<String, String> map) {
 		AttachmentVo vo = fs.getFile(map);
-		System.out.println(vo.getFilePath());
-		FileUnit.deleteFile(vo.getFilePath()+vo.getChangeName());
 		int result = fs.delete(map);
+		if(result == 1) {
+			FileUnit.deleteFile(vo.getFilePath()+vo.getChangeName());
+		}
 	}
 
 }
